tools/clinic-consult/pyproject.toml
[project]
name = "clinic-consult"
version = "0.1.0"
description = "Multi-platform messenger reservation bot for Korean dermatology clinics"
requires-python = ">=3.13"
dependencies = [
    "uiautomator2>=3.2",
    "anthropic>=0.42",
    "openai>=1.60",
    "httpx>=0.28",
    "pydantic>=2.9",
    "pydantic-settings>=2.7",
    "structlog>=24.4",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "apscheduler>=3.10",
    "jamo>=0.4",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "mypy>=1.14",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
